Update blocking-mode reboot test for Trixie (#21605)
What is the motivation for this PR?
Similar to #21394, the behavior of sudo reboot is changed with Trixie, in that the SSH connection may get disconnected before the sudo reboot command completes with an exit code. This will mean that Ansible will consider the host to be unreachable and throw an exception.

How did you do it?
The common reboot infra that we have runs sudo reboot in its own thread. That means that the HostUnreachable exception that Ansible raises will get raised in that thread, but won't affect the main thread; this, in a way, handles this particular behavior change. Therefore, for the blocking reboot test case, reuse that common infra when running on a Trixie image and we need to reboot the device (this probably can be done unconditionally, but make it conditional for now).

How did you verify/test it?
Tested with a subset of T0 KVM tests: https://elastictest.org/scheduler/testplan/6936fccd392767e9bf67ecb8

+++ tests/reboot/test_reboot_blocking_mode.py
@@ -1,5 +1,6 @@
 import pytest
 import re
+from tests.common.reboot import reboot
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.platform.processes_utils import wait_critical_processes
 
@@ -38,14 +39,18 @@ def restore_systemctl_reboot_and_reboot(duthost):
     execute_command(
         duthost,
         "sudo sed -i 's#/usr/local/bin/disabled_watchdogutil#/usr/local/bin/watchdogutil#g' /usr/local/bin/reboot")
-    execute_command(duthost, "sudo reboot")
+    debian_version = duthost.command("grep VERSION_CODENAME /etc/os-release")['stdout'].lower()
+    if "trixie" in debian_version:
+        reboot(duthost, safe_reboot=True)
+    else:
+        execute_command(duthost, "sudo reboot")
 
-    timeout = None
-    if duthost.is_supervisor_node():
-        timeout = 900
-    elif duthost.is_multi_asic:
-        timeout = 420
-    wait_critical_processes(duthost, timeout=timeout)
+        timeout = None
+        if duthost.is_supervisor_node():
+            timeout = 900
+        elif duthost.is_multi_asic:
+            timeout = 420
+        wait_critical_processes(duthost, timeout=timeout)
 
 
 def mock_reboot_config_file(duthost):
